pyproject.toml
[project]
name = "llm-judge-mvp"
version = "0.1.0"
description = ""
authors = [
    {name = "Arun",email = "arun@example.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)",
    "networkx (>=3.4,<4.0) ; python_version >= \"3.11\" and python_full_version != \"3.14.1\" and python_version < \"4.0\""
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "ruff (>=0.14.14,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
]

[tool.poetry]
name = "llm-judge-mvp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "llm_judge", from = "src" }]


[tool.poetry.group.dev.dependencies]
ruff = "^0.15.0"
mypy = "^1.19.1"
types-pyyaml = "^6.0.12.20250915"
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
datasets = "^4.8.4"


[tool.poetry.group.experiment.dependencies]
faiss-cpu = {version = ">=1.13.2,<2.0.0", markers = "python_version < \"3.15\""}
rank-bm25 = "^0.2.2"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["experiments/*"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"


[tool.mypy]
exclude = ["experiments/", "tests/"]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true

# We start pragmatic, not ultra-strict
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov=llm_judge --cov-report=term-missing --cov-fail-under=90 --strict-markers"
testpaths = ["tests"]
markers = [
    "nightly: long-running deterministic dataset runs (scheduled CI)",
]

omit = ["src/llm_judge/calibration/hallucination_graphs.py"]
[tool.coverage.run]
omit = [
  "src/llm_judge/calibration/hallucination_graphs.py",
  "src/llm_judge/logging.py",
  "src/llm_judge/eval/*",
  "src/llm_judge/judge_base.py",
  "src/llm_judge/benchmarks/funnel_analysis.py",
  "src/llm_judge/benchmarks/kg_extraction_experiment.py",
  "src/llm_judge/benchmarks/nli_gemini_benchmark.py",
  "src/llm_judge/benchmarks/nli_gemini_science_gate.py",
  "src/llm_judge/benchmarks/nli_science_gate.py",
  "src/llm_judge/benchmarks/qgqa_science_gate.py",
  "src/llm_judge/benchmarks/slm_gguf_science_gate.py",
  "src/llm_judge/benchmarks/slm_graphrag_science_gate.py",
  "src/llm_judge/benchmarks/slm_science_gate.py",
  "src/llm_judge/benchmarks/sweep_rag.py",
  "src/llm_judge/benchmarks/gate2_experiment.py",
  "src/llm_judge/benchmarks/graphrag_science_gate.py",
  "src/llm_judge/benchmarks/faithdial.py",
  "src/llm_judge/benchmarks/fever.py",
  "src/llm_judge/benchmarks/generate_ground_truth.py",
  "src/llm_judge/benchmarks/halueval.py",
  "src/llm_judge/benchmarks/ifeval.py",
  "src/llm_judge/benchmarks/jigsaw.py",
  "src/llm_judge/benchmarks/master_gt.py",
  "src/llm_judge/benchmarks/report.py",
  "src/llm_judge/benchmarks/run_all.py",
  "src/llm_judge/benchmarks/runner.py",
  "src/llm_judge/benchmarks/threshold_sweep.py",
  "src/llm_judge/benchmarks/toxigen.py",
  "src/llm_judge/benchmarks/verify_rag.py",
  "src/llm_judge/calibration/__main__.py",
  "src/llm_judge/datasets/__main__.py",
  "experiments/*"
]



